compare cards logic handled

+++ lib/interfaces/result.dart
@@ -1,8 +1,8 @@
 import 'package:cricket_card_game/enums.dart';
 
 class Result {
-  int activePlayerDamage;
-  int opponentPlayerDamage;
+  double activePlayerDamage;
+  double opponentPlayerDamage;
   ComparisonOutcome result;
   Result({
     required this.activePlayerDamage,
